client: network: elevate null handling from server post request to callee
Introduces a separate method to fetch the response data from the HTTP POST
requests to the game server. This allows a more flexible and generic usage of
the internal server client API. This is needed for a future commit, where a
response to a request does not contain any response data but the request state.

This also introduces its own exception for the existing player registration
class to be thrown if there was no player identifier returned by the server, as
this is required for the continued client initialization control flow.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>