client: map: implement receiving-side of map creation
Implements the creation of `GameMap` and its attributes' classes through
network library-provided data classes, which are needed when the game state is
polled and the new game state data is to be evaluated (i.e. if it changed and
if so, update the values accordingly).

This commit also dropped the `GameMapNode.placeTreasure()` method in favor of
setting the value of the `treasureState` when creating the class instance
through a more detailed constructor.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>