client: implement GameClientState and deserializing from network library
Implements the `GameClientState` class to hold the information about the
current game, including the GameMap and both players - even though the opponent
might not have joined yet.

This commit includes the conversion mapping from the network library-provided
`GameState`, which will allow the `GameClientState` to be updated by polling
the state from the game server and merging the instance's fields, which will be
implemented in a future commit.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>