client: implement game state polling method to retrieve the current game status
Implements a polling method in the `GameStateUpdater` for polling the current
game status, which is required to do for checking whether it's the
player/client's turn and to check if the map can be sent to the server.

This commit also implements the polling and map sending control flow in the
MainClient as a proof of concept, which should be built upon when the
MapGenerator is finished.

This also removes the now redundant example method (which was never needed in
the first place) to retrieve the current game state from the server.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>