client: map: implement debug rendering of player half map
Implements the debug rendering of `GameMap` class instances to a string by
displaying the 10 x 5 lattice, where each letter notates the terrain type of
the map node. For a map node containing a fort, this is the letter "F". For
other map nodes, this is "G" for grass terrain, "M" for mountain terrain and
"W" for water terrain.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>