client: generation: add fort structure generation
Implements an algorithm to generate a safe fort structure, which should act as
the safe starting point for the player's half map. The safety lies in the
design of the structure, as the fort can only be travelled to by one path and
can only be seen, if the player travels exactly to the fort's node. This is
because there is no possibility for an adjacent mountain map node.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>