client: validation: implement trivial half map validators
Implements the trivial half map validators to verify, that a generated half map
is compliant to the game's rules. The trivial validation cases are:

The half map...
- ...must contain exactly 50 map nodes,
- ...nodes must be within the x values 0 and 9,
- ...nodes must be within the y values 0 and 4,
- ...must have at least 24 map nodes with grass terrain [0],
- ...must have at least 5 map nodes with mountain terrain [1],
- ...must have at least 7 map nodes with water terrain [2], and
- ...must have at least 14 accessible border map nodes [3].

This implements a loop, which will generate maps until a valid map has been
generated to allow a fast, trivial random map generator to be implemented.

[0] 50 * 0.48 = 24
[1] 50 * 0.10 = 5
[2] 50 * 0.14 = 7
[3] ceil( (50 - 8 * 3) * 0.51 ) = 14, where 8 * 3 is the map without border.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>