client: validation: implement half map terrain reachability validator
Implements a validator to check whether the generated player's half map is free
of map nodes, which are accessible (i.e. the terrain is not water) but not
reachable (i.e. fully surrounded by inaccessible terrain type; an island).

The validator picks either the position of the fort or the origin position of
the map and floodfills the existing `GameMap` instance by walking to each
reachable and neighboring map node and marking it as visited. This is done in a
depth-first search approach. Afterwards, the set of all accessible map nodes
are verified to be a subset of the set of visited map nodes.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>