client: generation: implement simple random map generator
Implements a random `MapGenerator`, which generates a pre-defined fort
structure and generates a pseudorandom terrain distribution around the existing
fort structure.

To ensure that the required amount of map nodes per terrain are reached, the
required terrain types are pre-queued, where the remaining possible terrains
are filled with mountain terrain types to maximize mountains in the generated
`GameMap`. Afterwards, the terrain queue is pseudo-randomized and used to
generate the remaining `GameMapNode`s for the `GameMap`.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>

Date:      Sun Dec 1 17:08:01 2024 +0100