client: validation: fix border accessibility validator
Implements the border accessibility validator as a validation rule, where the
border nodes in each border direction must contain at least 51 % accessible
nodes to be accepted as a correctly generated `GameMap`.

This fixes the behavior of the border accessibility validator implemented in
af25030, which validates whether the whole player half map's border contains at
least 51 % (i.e. 14) accessible map nodes. This is wrong and will result in the
server rejecting the sent map.

Fixes: af25030 ("client: validation: implement trivial half map validators")
Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>