client: player: add log statements on player update changes
Implements log statements for changes to the player's position and whether they
have found their treasure on updates to a `Player` instance.

To have a user-visible handle for a player, a player handle is constructed from
the player's uaccount name and the first five letters of the player's unique
identifier.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>