client: map: introduce `Path` and `Step` classes for `PathFinder`
Implements the `Path` and `Step` classes, which are to be used for future
implementations of the `PathFinder` interface's `findPath` function. This
provides a abstract protocol, which allows instances of `Path` to be created
independent of a concrete `GameMap` instance, e.g. just a collection of
relevant `GameMapNode`s.

The `Path` class provides a ready-made `intoMapDirections` method, which allows
the `Path` to be mapped or "projected" onto a concrete `GameMap` instance by
expanding the path's vertices into a list of `Step`s.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>