client: validation: refactor game map validator to use notification pattern
Implements the notification pattern in the `HalfMapValidator` class to allow
reporting of individual violation of business rules when validating a player's
half map during map generation.

This modifies the `GameMapValidationRule` API by adding to a `Notification<>`
instance instead of returning only a binary boolean value to allow more
clarification of what has gone wrong. The violations are listed in a
bullet-point list when logged as a string.

For most existent checks the conditions are split to accomodate a more granular
description of the business rule violation and are negated as the former
`validate(GameMap)` API returned `true` in case of success, but the
notification entry addition does only run for the `false` states now.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>