client: map: document and rename `resetVisibility` to `resetIntelligence`
Renames the class method `GameMapNode::resetVisibility` to
`GameMapNode::resetIntelligence` as it reflects the intention of the method's
functionality better: The `FortState` and `TreasureState` are the intelligence,
i.e. knowledge, we have on it. It had nothing to do with the visibility of it.

Since the intricacies of the extended mechanism is a little bit unclear,
document the reason behind this method at its definition and only invocation.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>