client: test: map: implement test cases for `GameMap::get{Player,Enemy}MapNodes`
Implement test cases for the `GameMap::get{Player,Enemy}MapNodes` methods and
if they return the proper half map's nodes with respect to the position of the
player's or enemy's fort.

This introduces the `GameMap(Map<Position, GameMapNode>)` constructor and the
`GameMapNode::placePlayerFort` method to allow proper access to the underlying
classes without losing encapsulation.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>