client: test: generation: implement test case for `MapGenerator`
Since all implementations of `GameMapValidationRule` are backed by an extensive
enough test suite to verify their correctness, implement a simple timeout-based
test case for the `MapGenerator::generateUntilValid` method, which makes sure
that the generation of a `GameMap` does not cause an infinite loop caused by
any changes in the validators.

This is reliabe enough for now, but should be extended with other test cases.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>