client: map: fix calculation of `PositionArea` corner positions
The corner calculation for `PositionArea` had two flaws:
- the calculation including the width were wrong, and
- the `lowerRight` used the `width` to calculate the `y` coordinate

The implementation of `PositionArea`'s corner position calculation methods were
introduced for test cases for the `GameMapTerrainReachabilityValidator` to
construct a `GameMap` with a inaccessible border, but leaving the corners of
the border as grass fields, i.e. having reachable, inaccessible grass fields in
the corners.

This has gone unnoticed, because of the sleepiness of the programmer and since
the product of these methods were only used in the method `isCorner` and so at
least one corner node could be successfully generated.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>