client: map: fix middle point calculation of `PositionArea`
Fix the wrong calculation from calculating the half of the difference between
`width`/`height` and `x`/`y` to calculating the half of the `width`/`height`
and adding the starting `x`/`y` value.

This has gone unnoticed until now since the only callee was `GameMap`, which
always has `(0, 0)` as the value of `(x, y)`.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>