client: map: simplify retrieval of enemy's half map nodes
Splits the previously introduced `PositionArea::intoHalfStream` into a current
and other half area part, so that it can be neatly reused for the retrieval of
the enemy's half map nodes.

This replaces the more tedious former retrieval by calculating the reflective
point on the other side of the `GameMap`. This has been a hotspot for easy bugs
and errors, so this code should be much safer and more readable.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>