client: main: revert marking the client to win the game if at enemy fort
In hindsight, it is better to rely soley on the game server's state whether the
player has been marked as having won or not. If the player is on the enemy fort
and the game server hasn't set the player to have won, it's much more likely
that the client has constructed a invalid game state.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>