client: test: main: implement perfect knowledge test case for `GameClient`
Implements a test case against the `GameClient::run` method to verify that the
`GameClient` can successfully pass all stages with perfect knowledge of a
simple, empty (i.e. only grass fields) map.

To assert the game client's win statement, the `GameClientState` needs to be
exposed to the `GameClient` class users, which is needed for the transition to
the MVC architecture later on anyway.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>