client: search: implement initial version of fort estimator
Implements the `FortEstimator` class, which in its initial version allows a
user to retrieve a set of possible positions, where the player's initial
starting position could be even though there are at least two positions or at
most four positions missing from the player's position history.

The initial starting position can be determined by determining which positions
could have been reached from the first known reliable position in a specified
amount of time.

This is mainly used to determine a set of possible positions for the enemy's
fort, since the first eight turns in the game the client does only get random
enemy positions on game state updates. Therefore, the first enemy location,
which is followed by an adjacent position (and so on), is assumed to be the
first reliable location.

This implementation can be improved by using other knowledge:
- using the direction of the position history as an indicator that the possible
  position is more likely to be in the reverse or perpendicular direction than
  where the first few positions are pointing to, and
- using the whole position history which positions have been traversed by the
  player, since the player is unlikely to head back to their own fort.

Signed-off-by: Daniel Kral <a11908284@unet.univie.ac.at>